+++ src/io/mod.rs
@@ -1,1 +1,1 @@
-mod io_csv;
+pub mod csv;
