+++ tests/lib.rs
@@ -2,3 +2,8 @@
 extern crate rulinalg;
 
 pub mod mat;
+
+#[cfg(feature = "io")]
+pub mod io {
+    mod csv;
+}
